NAS-107393 Move items to new section
