+++ src/app/pages/system/general-settings/general-settings.component.scss
@@ -103,3 +103,153 @@
     }
 
 }
+
+// .sys-image { not working here
+//     display: flex;
+//     justify-content: center;
+//     align-items: flex-start;
+//     margin-bottom: 1rem;
+
+//     div {
+//     display: flex;
+//     justify-content: center;
+//     }
+// }
+
+.lowerme {
+    margin-top: 60px !important;
+}
+
+#category_form-select-spinner,
+app-proactive #name_form-input-spinner,
+app-proactive #secondary_name_form-input-spinner {
+    position: absolute;
+    top: -20px;
+    right: 45%;
+}
+
+.button-group button {
+    margin-right: 8px;
+}
+
+.button-group.btn-fixed {
+    position: absolute;
+    bottom: 25px;
+    left: 25px;
+
+    @media screen and (max-width: 789px) {
+        position: relative;
+        bottom: 0;
+        left: 0;
+    }
+}
+
+//   app-sys-info {
+    .label {
+      opacity: .8;
+      margin-right: 4px;
+    }
+    .sys-info-wrapper {
+      margin-top: 5px;
+      margin-bottom: 1rem;
+    }
+    .sys-info-line {
+      margin-bottom: 5px;
+    }
+//   }
+
+//   app-sys-image {
+    img {
+      max-width: 90%;
+      height: auto;
+      max-height: 325px;
+
+      @media screen and (max-width: 789px) {
+        max-height: 250px;
+        margin-bottom: 1rem;
+      }
+    }
+//   }
+
+  app-production-status {
+    .mat-card {
+        top: 25px;
+        background-color: rgba(127,127,127,0.1);
+        width: 400px;
+        height: 100px;
+        margin: 0 !important;
+    }
+    .mat-card-actions {
+      top: 25px;
+      position: absolute;
+      left: 240px;
+      width: unset;
+    }
+
+    @media screen and (max-width: 599px) {
+      .mat-card {
+        top: 25px;
+        width: auto;
+      }
+    }
+
+    @media screen and (max-width: 460px) {
+      .mat-card {
+        top: 25px;
+        height: auto;
+      }
+      .mat-card-actions {
+        position: relative;
+        top: unset;
+        left: unset;
+      }
+    }
+  }
+
+  app-proactive {
+    .mat-card {
+      top: 37px;
+      background-color: rgba(127,127,127,0.1);
+      margin: 0 15px !important;
+    }
+    .mat-card-actions {
+      height: 0px;
+      top: -45px;
+      position: relative;
+      left: 230px;
+    }
+    #enabled {
+      position: relative;
+      top: 5px;
+    }
+
+    #TN_proactive_instructions {
+      font-size: 120%;
+      margin-right: 10px;
+      margin-bottom: 20px;
+
+    }
+
+    @media screen and (max-width: 599px) {
+      .mat-card {
+        top: -35px;
+      }
+    }
+
+    @media screen and (max-width: 460px) {
+      .mat-card-actions {
+        top: 25px;
+        height: auto;
+        position: relative;
+        left: unset;
+        height: auto;
+        top: unset;
+      }
+    }
+  }
+
+  app-entity-dialog #tn_core_upgrade_license_message {
+    margin-bottom: 8px;
+  }
+
+// End support section //////////////
