+++ src/app/pages/system/general-settings/ntpservers/ntpserver-form/ntpserver-form.component.ts
@@ -3,9 +3,9 @@ import { ActivatedRoute, Router } from '@angular/router';
 import { Validators } from "@angular/forms";
 
 import { helptext_system_ntpservers as helptext } from 'app/helptext/system/ntpservers';
-import { WebSocketService } from '../../../../services';
+import { WebSocketService } from '../../../../../services';
 import { ModalService } from 'app/services/modal.service';
-import { FieldConfig } from '../../../common/entity/entity-form/models/field-config.interface';
+import { FieldConfig } from '../../../../common/entity/entity-form/models/field-config.interface';
 import { FieldSet } from 'app/pages/common/entity/entity-form/models/fieldset.interface';
 import { greaterThan } from "app/pages/common/entity/entity-form/validators/compare-validation";
