+++ src/app/pages/system/general-settings/support/eula/eula.component.html
