+++ src/app/pages/system/general-settings/support/support-licensed/support-form-licensed.component.ts
@@ -10,7 +10,7 @@ import { EntityJobComponent } from 'app/pages/common/entity/entity-job/entity-jo
 import { FieldConfig } from 'app/pages/common/entity/entity-form/models/field-config.interface';
 import { FieldSet } from 'app/pages/common/entity/entity-form/models/fieldset.interface';
 import { helptext_system_support as helptext } from 'app/helptext/system/support';
-import { ModalService } from '../../../../services/modal.service';
+import { ModalService } from '../../../../../services/modal.service';
 
 @Component({
   selector: 'app-support-form-licensed',
