+++ src/app/pages/system/general-settings/support/sys-image/sys-image.component.ts
@@ -2,7 +2,8 @@ import { Component, Input } from '@angular/core';
 
 @Component({
   selector: 'app-sys-image',
-  templateUrl: './sys-image.component.html'
+  templateUrl: './sys-image.component.html',
+  styleUrls: ['../../general-settings.component.scss']
 })
 export class SysImageComponent  {
   @Input() product_image: string;
