+++ src/app/pages/system/general-settings/support/sys-info/sys-info.component.html
