--- src/app/pages/system/general/general.component.css
@@ -1,35 +1,0 @@
-.mat-fab-bottom-right-1 {
-  top: auto !important;
-  right: 30px !important;
-  bottom: 48px !important;
-  left: auto !important;
-  position: fixed !important;
-  z-index: 2;
-}
-
-.mat-fab-bottom-right-2 {
-  top: auto !important;
-  right: 90px !important;
-  bottom: 48px !important;
-  left: auto !important;
-  position: fixed !important;
-  z-index: 2;
-}
-
-.mat-fab-bottom-right-3 {
-  top: auto !important;
-  right: 150px !important;
-  bottom: 48px !important;
-  left: auto !important;
-  position: fixed !important;
-  z-index: 2;
-}
-
-.mat-fab-bottom-right-4 {
-  top: auto !important;
-  right: 210px !important;
-  bottom: 48px !important;
-  left: auto !important;
-  position: fixed !important;
-  z-index: 2;
-}
