--- src/app/pages/system/general/index.ts
@@ -1,4 +1,0 @@
-// export {ConfigResetComponent} from './config-reset/';
-// export {ConfigSaveComponent} from './config-save/';
-// export {ConfigUploadComponent} from './config-upload/';
-//export {GeneralComponent} from './general.component';
