--- src/app/pages/system/ntpservers/ntpserver-list/index.ts
@@ -1,1 +1,0 @@
-export {NTPServerListComponent} from './ntpserver-list.component';
