--- src/app/pages/system/support/license/license.component.html
@@ -1,1 +1,0 @@
-<p>license works!</p>
