+++ src/assets/styles/tn-styles.scss
@@ -1126,296 +1126,6 @@ $primary-dark: darken( map-get($md-primary, 500), 8% );
     top: -18px;
   }
 
-  // Support Section ////////////////////////
-  app-support {
-    // .mat-card {
-    //   min-width: 320px;
-    // }
-    // #top-row {
-    //   margin: 0 15px;
-    // }
-
-    .sys-image {
-      display: flex;
-      justify-content: center;
-      align-items: flex-start;
-      margin-bottom: 1rem;
-
-      div {
-        display: flex;
-        justify-content: center;
-      }
-    }
-
-    .lowerme {
-      margin-top: 60px !important;
-    }
-
-    .support-title, #FN_col1, #FN_col2, #TN_col1, #TN_proactive_section_title {
-      font-size: 180%;
-      margin-bottom: 15px;
-
-      & .material-icons {
-        margin-right: 10px;
-        margin-left: -5px;
-        position: relative;
-        top: 10px;
-        font-size: 36px;
-      }
-    }
-
-    #category_form-select-spinner,
-    app-proactive #name_form-input-spinner,
-    app-proactive #secondary_name_form-input-spinner {
-      position: absolute;
-      top: -20px;
-      right: 45%;
-    }
-
-    .button-group.btn-fixed {
-      position: absolute;
-      bottom: 25px;
-      left: 25px;
-
-      button {
-        margin-right: 8px;
-      }
-    }
-
-    @media screen and (max-width: 789px) {
-      .button-group.btn-fixed {
-        position: relative;
-        bottom: 0;
-        left: 0;
-      }
-    }
-
-    @media screen and (max-width: 1025px) {
-      #fn-spacer {
-        height: 80px;
-      }
-    }
-
-    @media screen and (max-width: 400px) {
-      app-sys-image img {
-        max-width: 80%;
-      }
-      .mat-card {
-        margin-left: -10px;
-      }
-    }
-
-    @media screen and (max-width: 350px) {
-      .mat-card {
-        margin-left: -20px;
-      }
-    }
-
-    .links-wrapper {
-      margin-bottom: 1rem;
-    }
-    .links {
-      margin-bottom: 5px;
-
-      a {
-        color: var(--primary);
-      }
-    }
-    
-
-  }
-
-  app-sys-info {
-    .label {
-      opacity: .8;
-      margin-right: 4px;
-    }
-    .sys-info-wrapper {
-      margin-top: 5px;
-      margin-bottom: 1rem;
-    }
-    .sys-info-line {
-      margin-bottom: 5px;
-    }
-  }
-
-  app-tn-sys-info {
-    .tn-sys-info-line {
-      margin-top: 10px;
-
-      h4 {
-        margin-bottom: 3px;
-        display: inline-block;
-      }
-    }
-    #update-license-btn {
-      background-color: var(--primary);
-      color: white;
-      top: 13px;
-    }
-
-    #tn-sys-col1 {
-      min-width: 184px;
-    }
-  }
-
-  app-sys-image {
-    img {
-      max-width: 90%;
-      height: auto;
-      max-height: 325px;
-
-      @media screen and (max-width: 789px) {
-        max-height: 250px;
-        margin-bottom: 1rem;
-      }
-
-    }
-  }
-
-  @media screen and (min-width: 600px) and (max-width: 749px)  {
-    .tn-sys-image {
-      position: relative;
-      left: 35px;
-    }
-  }
-
-  app-production-status {
-    .mat-card {
-        top: 25px;
-        background-color: rgba(127,127,127,0.1);
-        width: 400px;
-        height: 100px;
-        margin: 0 !important;
-    }
-    .mat-card-actions {
-      top: 25px;
-      position: absolute;
-      left: 240px;
-      width: unset;
-    }
-
-    @media screen and (max-width: 599px) {
-      .mat-card {
-        top: 25px;
-        width: auto;
-      }
-    }
-
-    @media screen and (max-width: 460px) {
-      .mat-card {
-        top: 25px;
-        height: auto;
-      }
-      .mat-card-actions {
-        position: relative;
-        top: unset;
-        left: unset;
-      }
-    }
-  }
-
-  app-proactive {
-    .mat-card {
-      top: 37px;
-      background-color: rgba(127,127,127,0.1);
-      margin: 0 15px !important;
-    }
-    .mat-card-actions {
-      height: 0px;
-      top: -45px;
-      position: relative;
-      left: 230px;
-    }
-    #enabled {
-      position: relative;
-      top: 5px;
-    }
-
-    #TN_proactive_instructions {
-      font-size: 120%;
-      margin-right: 10px;
-      margin-bottom: 20px;
-
-    }
-
-    @media screen and (max-width: 599px) {
-      .mat-card {
-        top: -35px;
-      }
-    }
-
-    @media screen and (max-width: 460px) {
-      .mat-card-actions {
-        top: 25px;
-        height: auto;
-        position: relative;
-        left: unset;
-        height: auto;
-        top: unset;
-      }
-    }
-  }
-
-  app-fn-support {
-    .mat-card {
-      background-color: rgba(127,127,127,0.1);
-      margin: 30px 15px 5px 15px !important;
-    }
-    .mat-card-actions {
-      top: -40px;
-      position: relative;
-      width: 300px;
-    }
-    form {
-      margin-bottom: -40px;
-    }
-
-    @media screen and (max-width: 599px) {
-      .mat-card-actions {
-        top: 0;
-      }
-      .lowerme {
-        margin-top: 0 !important;
-      }
-      form {
-        margin-bottom: 4px;
-      }
-    }
-
-  }
-
-  app-entity-dialog #tn_core_upgrade_license_message {
-    margin-bottom: 8px;
-  }
-
-  app-tn-support {
-    .mat-card {
-      background-color: rgba(127,127,127,0.1);
-      margin: 50px 15px 5px 15px !important;
-    }
-    .mat-card-actions {
-      height: 0px;
-      top: -55px;
-      position: relative;
-      width: 400px;
-    }
-
-    @media screen and (max-width: 599px) {
-      .mat-card-actions {
-        top: 0;
-        height: 25px !important;
-      }
-      .lowerme {
-        margin-top: 0 !important;
-      }
-      .mat-card {
-        top: -15px;
-      }
-    }
-  }
-// End support section //////////////
-
   entity-form form-combobox mat-form-field {
     width: 50%;
     min-width: 180px;
@@ -1484,11 +1194,6 @@ $primary-dark: darken( map-get($md-primary, 500), 8% );
   *::-webkit-scrollbar-thumb{
     background-color: rgba(127,127,127,0.5);
     border-radius:4px;
-  }
-
-
-  app-support a {
-    text-decoration: underline;
   }
 
   app-entity-dialog #pool_processes p,
@@ -1834,6 +1539,29 @@ $primary-dark: darken( map-get($md-primary, 500), 8% );
     .lowerme-guiform {
       margin-top: 32px !important;
     }
+
+    app-support .links-wrapper {
+      margin-bottom: 1rem;
+    }
+    app-support .links {
+      margin-bottom: 5px;
+      a {
+        color: var(--primary);
+        text-decoration: underline;
+      }
+    }
+
+    .sys-image {
+      display: flex;
+      justify-content: center;
+      align-items: flex-start;
+      margin-bottom: 1rem;
+  
+      div {
+      display: flex;
+      justify-content: center;
+      }
+  }
 
 
 } // end of ix theme
